% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test.multiple}
\alias{test.multiple}
\title{Conduct multiple tests}
\usage{
test.multiple(Y, X, map, w.type = NULL, spec = 1, min = 100,
  steps = 20)
}
\arguments{
\item{Y}{exon expression\strong{:}
matrix with \eqn{n} rows (samples) and \eqn{p} columns (exons)}

\item{X}{SNP genotype\strong{:}
matrix with \eqn{n} rows (samples) and \eqn{q} columns (SNPs)}

\item{map}{list with names "genes", "exons", and "snps"
(output from \code{\link{map.genes}}, \code{\link{map.exons}},
and \code{\link{map.snps}})}

\item{w.type}{string\strong{:}
if \code{W=NULL} and w.type is a given string, take the type: "cov" is the only one allowed. 
Then the inner product of the indep data \code{X} is taken}

\item{spec}{number of cores\strong{:}
positive integer}

\item{min}{minim chunk size\strong{:}
positive integer}

\item{steps}{number of iteration chunks\strong{:}
positive integer}
}
\description{
This function tests for alternative splicing.
}
\details{
Automatic adjustment of the number of permutations
such that Bonferroni-significant p-values are possible.
}
\examples{
NA

}
